+++ src/Cwaff-Guns/SeltzerPelter.cs
@@ -19,7 +19,7 @@ public class SeltzerPelter : CwaffGun
           .AddToShop(ItemBuilder.ShopType.Goopton)
           .AssignGun(out Gun gun)
           .InitProjectile(GunData.New(clipSize: 1, cooldown: 0.5f, shootStyle: ShootStyle.SemiAutomatic, customClip: true, preventOrbiting: true,
-            damage: 16.0f, speed: 30.0f, force: 75.0f, range: 999.0f, sprite: "can_projectile", fps: 1, anchor: Anchor.MiddleCenter, // 1 FPS minimum, stop animator manually later
+            damage: 16.0f, speed: 30.0f, force: 45.0f, range: 999.0f, sprite: "can_projectile", fps: 1, anchor: Anchor.MiddleCenter, // 1 FPS minimum, stop animator manually later
             anchorsChangeColliders: false, overrideColliderPixelSizes: new IntVector2(2, 2))) // prevent uneven colliders from glitching into walls
           .Attach<BounceProjModifier>(bounce => {
               bounce.numberOfBounces      = 9999;